[action] [PR:23106] [logrotate] Add archived file clean up on log rotation (#1329)
<!--
 Please make sure you've read and understood our contributing guidelines:
 https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

 failure_prs.log skip_prs.log Make sure all your commits include a signature generated with `git commit -s` **

 If this is a bug fix, make sure your description includes "fixes #xxxx", or
 "closes #xxxx" or "resolves #xxxx"

 Please provide the following information:
-->

#### Why I did it
To fix https://github.com/sonic-net/sonic-buildimage/issues/23053 ,

In a corner case, the log partition(/var/log) is with 100% utilization, and the logrotate will never release space, and no new logs can be write .
Conditions that need to be met:

1. /var/log partition is 100% utilized.
2. The utilizations are on origin log files(e.g. syslog, auth.log, swss.asic0.rec), and no atchived log files(e.g. syslog.1, syslog.2.gz, sairedis.asic0.rec.1)
3. The logrotate will never free/delete log files on this scenario

Add archived file clean up on log rotation, to make sure the log files can be rotated and deleted.
The logic will only be triggered on syslog(moved to the last one in the log list) rotation, to avoid useless scanning.

##### Work item tracking
- Microsoft ADO **(number only)**: 33258706

#### How I did it
Add archived file clean up on log rotation, to make sure the log files can be rotated and deleted.
The logic will only be triggered on syslog rotation, to avoid useless scanning.
#### How to verify it
Verified:
1. Logrotate can deleted archived files when disk is full.
2. Logrotate works well on normal state.
<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 202205
- [ ] 202211
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] 202505

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [ ] <!-- image version 1 -->
- [ ] <!-- image version 2 -->

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)