Fix Debian repos used for Bullseye-based containers
When building a Bullseye-based container with MIRROR_SNAPSHOT=y, the
repos used will point to the main Debian repo, but Bullseye packages
have been moved to the archive repo. Fix this issue by overriding the
`DEFAULT_MIRROR_URLS` variable.

Note that bullseye-security is still in the main Debian repo, and is not
yet on the archive repo.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>